+++ javai-tagging/src/test/java/dev/xtrafe/javai/tagging/TestTextedNoteRepository.java
@@ -1,0 +1,7 @@
+package dev.xtrafe.javai.tagging;
+
+import dev.xtrafe.javai.persistence.JavAIRepository;
+
+/** Ordinary realized repository for {@link TestTextedNote} -- same pattern as {@link TestTextedThingRepository}. */
+public interface TestTextedNoteRepository extends JavAIRepository<TestTextedNote> {
+}